windows_build_wheels: Move util functions into dedicated module
